splitting out context
