+++ Config.cmake/fmixml.cmake
@@ -150,6 +150,7 @@ set(FMIXMLHEADERS
     include/FMI3/fmi3_xml_variable.h
     src/FMI3/fmi3_xml_variable_impl.h
     src/FMI3/fmi3_xml_parser_lists.h
+    src/FMI3/fmi3_xml_parser_context_impl.h
     src/
  )
